+++ hutool-poi/src/test/java/cn/hutool/poi/excel/ExcelReadTest.java
@@ -245,9 +245,9 @@ public class ExcelReadTest {
 		Assert.assertEquals("", objects.get(2));
 	}
 
-	@Test(expected = NullPointerException.class)
-	@Ignore
+	@Test
 	public void readColumnNPETest() {
+		// https://github.com/dromara/hutool/pull/2234
 		ExcelReader reader = ExcelUtil.getReader(ResourceUtil.getStream("read_row_npe.xlsx"));
 		reader.readColumn(0, 1);
 	}