+++ hutool-core/src/test/java/cn/hutool/core/lang/ConsoleTest.java
@@ -11,14 +11,14 @@ import cn.hutool.core.thread.ThreadUtil;
  *
  */
 public class ConsoleTest {
-	
+
 	@Test
 	public void logTest(){
 		Console.log();
-		
+
 		String[] a = {"abc", "bcd", "def"};
 		Console.log(a);
-		
+
 		Console.log("This is Console log for {}.", "test");
 	}
 
@@ -27,12 +27,12 @@ public class ConsoleTest {
 		Console.log("a", "b", "c");
 		Console.log((Object) "a", "b", "c");
 	}
-	
+
 	@Test
 	public void printTest(){
 		String[] a = {"abc", "bcd", "def"};
 		Console.print(a);
-		
+
 		Console.log("This is Console print for {}.", "test");
 	}
 
@@ -41,14 +41,14 @@ public class ConsoleTest {
 		Console.print("a", "b", "c");
 		Console.print((Object) "a", "b", "c");
 	}
-	
+
 	@Test
 	public void errorTest(){
 		Console.error();
-		
+
 		String[] a = {"abc", "bcd", "def"};
 		Console.error(a);
-		
+
 		Console.error("This is Console error for {}.", "test");
 	}
 
@@ -57,7 +57,7 @@ public class ConsoleTest {
 		Console.error("a", "b", "c");
 		Console.error((Object) "a", "b", "c");
 	}
-	
+
 	@Test
 	@Ignore
 	public void inputTest() {
@@ -65,7 +65,7 @@ public class ConsoleTest {
 		String input = Console.input();
 		Console.log(input);
 	}
-	
+
 	@Test
 	@Ignore
 	public void printProgressTest() {